Moved window manager related files to wm folder
