--- source/display/WindowTree.cpp
@@ -1,1 +1,0 @@
-#include "display/WindowTree.hpp"
