+++ source/wm/WindowTree.cpp
@@ -1,0 +1,1 @@
+#include "wm/WindowTree.hpp"
